requirements.txt
absl-py==0.11.0
addict==2.4.0
albumentations==1.0.3
argparse==1.4.0
atomicwrites==1.4.0
attrs==21.2.0
batchgenerators==0.24
bitarray==1.6.1
blas==1
ca-certificates==2020.12.5
cached-property==1.3.0
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.3.9
conda-pack==0.4.0
connected-components-3d==1.7.0
cryptography==2.9.2
cycler==0.10.0
cython==0.26
dataclasses==0.8
deap==1.3.1
decorator==4.4.1
deprecated==1.2.4
dicom2nifti==1.1.11
easydict==1.9
einops==0.3.0
entrypoints==0.3
freetype==2.10.4
future==0.18.2
git==2.11.1
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
gwdt==0.1.0
h5py==3.1.0
hdf5==1.10.6
icu==67.1
idna==2.9
imageio==2.9.0
importlib-metadata==3.1.1
iniconfig==1.1.1
joblib==0.11
jpeg==9d
kiwisolver==1.3.1
krb5==1.17.2
libblas==3.9.0
libcblas==3.9.0
libclang==10.0.1
libcurl==7.71.1
liblapack==3.9.0
libpng==1.6.37
libssh2==1.9.0
libtiff==0.4.2
linecache2==1.0.0
lz4-c==1.9.2
m2w64-gcc-libgfortran==5.3.0
m2w64-gcc-libs==5.3.0
m2w64-gcc-libs-core==5.3.0
m2w64-gmp==6.1.0
m2w64-libwinpthread-git==5.0.0.4634.697f757
markdown==3.3.3
matlabengineforpython==R2018a
matplotlib==3.3.3
matplotlib-base==3.3.3
medpy==0.4.0
mkl==2017.0.3
mmcv==1.3.9
msys2-conda-epoch==20160418
networkx==2.5
nibabel==3.2.1
nnunet==1.7.0
numpy==1.16.0
oauthlib==3.1.0
olefile==0.44
opencv-contrib-python==3.4.3.18
opencv-python==4.4.0.46
opencv-python-headless==4.5.3.56
openssl==1.1.1k
packaging==20.7
pandas==1.1.4
pillow==6.2.2
pip==20.2.4
pluggy==0.13.1
protobuf==3.14.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2
pycparser==2.2
pydicom==2.1.1
pylibtiff==0.4.2
pyparsing==2.2.0
pyqt==5.12.3
pyqt5-sip==4.19.18
pyqtchart==5.12
pyqtwebengine==5.12.1
pyreadline==2.1
pytest==6.2.4
python==3.6.12
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-utils==2.4.0
python_abi==3.6
pytz==2020.4
pywavelets==1.1.1
pyyaml==5.3.1
qt==5.12.9
regex==2021.7.6
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-fmm==2021.2.2
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.3
self-attention-cv==1.1.0
setuptools==50.3.1
simpleitk==2.0.1
six==1.10.0
sklearn
sklearn-contrib-lightning==0.5.0
sklearn-contrib-py-earth==0.1.0
sklearn-crfsuite==0.3.6
sklearn-genetic==0.3.0
sqlite==3.33.0
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardx==2.1
thop==0.0.31-2005241907
threadpoolctl==2.1.0
tifffile==2020.9.3
timm==0.4.12
tk==8.6.10
toml==0.10.2
torch==1.10.2
torch-summary==1.4.5
torchaudio==0.7.0
torchstat==0.0.7
torchsummary==1.5.1
torchvision==0.8.1+cu110
tornado==4.5.2
tqdm==4.15.0
traceback2==1.4.0
typing-extensions==3.7.4.3
unittest2==1.1.0
urllib3==1.26.2
vc==14.1
vs2015_runtime==14.16.27012
werkzeug==1.0.1
wheel==0.35.1
wincertstore==0.2
wrapt==1.12.1
xlsxwriter==0.9.8
xz==5.2.5
yacs==0.1.8
yaml==0.1.6
yapf==0.31.0
zipp==3.4.0

